Separate code into partial classes
